pyproject.toml
[project]
name = "garf"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = "===3.12.3"
dependencies = [
    "accelerate>=1.12.0",
    "addict>=2.4.0",
    "diffusers>=0.32.2",
    "h5py>=3.13.0",
    "hydra-core>=1.3.2",
    "lightning>=2.5.1",
    "matplotlib>=3.10.8",
    "mitsuba>=3.7.1",
    "peft>=0.18.0",
    "rich==13.9.4",
    "scipy>=1.15.2",
    "spconv-cu126>=2.3.8",
    "timm>=1.0.22",
    "torch==2.8.0",
    "torchvision>=0.23.0",
    "transformers>=4.50.2",
    "trimesh>=4.6.6",
    "typeguard>=4.4.2",
    "wandb>=0.19.9",
]

[dependency-groups]
dev = [
    "ipykernel>=6.29.5",
    "pre-commit>=4.2.0",
    "pytest>=8.3.5",
    "rerun-sdk>=0.22.1",
    "ruff>=0.11.2",
]


[project.optional-dependencies]
post = [
    "flash-attn",
    "pytorch3d",
    "torch-scatter==2.1.2+pt28cu128",
    "torch-cluster==1.6.3+pt28cu128",
]

[tool.uv]
environments = ["sys_platform == 'linux'"]
no-build-isolation-package = [
    "flash-attn",
    "pytorch3d",
    "torch-scatter",
    "torch-cluster",
]
find-links = [
    "https://data.pyg.org/whl/torch-2.8.0+cu128.html",
    "https://nvidia-kaolin.s3.us-east-2.amazonaws.com/torch-2.8.0_cu128.html",
]

[tool.uv.sources]
source = { workspace = true }
pytorch3d = { url = "https://github.com/MiroPsota/torch_packages_builder/releases/download/pytorch3d-0.7.8%2B5043d15/pytorch3d-0.7.8+5043d15pt2.8.0cu128-cp312-cp312-linux_x86_64.whl" }
flash-attn = { url = "https://github.com/Dao-AILab/flash-attention/releases/download/v2.8.3/flash_attn-2.8.3+cu12torch2.8cxx11abiTRUE-cp312-cp312-linux_x86_64.whl" }
craftsman = { path = "third_party/craftsman", editable = true }
torch = { index = "pytorch" }
torchvision = { index = "pytorch" }

[[tool.uv.index]]
name = "pytorch"
url = " https://download.pytorch.org/whl/cu128"
explicit = true

[tool.pytest.ini_options]
markers = ["manual: should be run manually."]
testpaths = ["source"]
